+++ src/WebhookConfig.php
@@ -5,6 +5,8 @@ namespace Spatie\WebhookClient;
 use Spatie\WebhookClient\Exceptions\InvalidConfig;
 use Spatie\WebhookClient\SignatureValidator\SignatureValidator;
 use Spatie\WebhookClient\WebhookProfile\WebhookProfile;
+use Spatie\WebhookClient\WebhookResponse\DefaultRespondsTo;
+use Spatie\WebhookClient\WebhookResponse\RespondsToWebhook;
 
 class WebhookConfig
 {
@@ -18,9 +20,11 @@ class WebhookConfig
 
     public WebhookProfile $webhookProfile;
 
+    public RespondsToWebhook $webhookResponse;
+
     public string $webhookModel;
 
-    public ProcessWebhookJob $processWebhookJob;
+    public string $processWebhookJobClass;
 
     public function __construct(array $properties)
     {
@@ -40,11 +44,17 @@ class WebhookConfig
         }
         $this->webhookProfile = app($properties['webhook_profile']);
 
+        $webhookResponseClass = $properties['webhook_response'] ?? DefaultRespondsTo::class;
+        if (! is_subclass_of($webhookResponseClass, RespondsToWebhook::class)) {
+            throw InvalidConfig::invalidWebhookResponse($webhookResponseClass);
+        }
+        $this->webhookResponse = app($webhookResponseClass);
+
         $this->webhookModel = $properties['webhook_model'];
 
         if (! is_subclass_of($properties['process_webhook_job'], ProcessWebhookJob::class)) {
             throw InvalidConfig::invalidProcessWebhookJob($properties['process_webhook_job']);
         }
-        $this->processWebhookJob = app($properties['process_webhook_job']);
+        $this->processWebhookJobClass = $properties['process_webhook_job'];
     }
 }