+++ tests/TestClasses/CustomRespondsToWebhook.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Spatie\WebhookClient\Tests\TestClasses;
+
+use Illuminate\Http\Request;
+use Spatie\WebhookClient\WebhookConfig;
+use Spatie\WebhookClient\WebhookResponse\RespondsToWebhook;
+use Symfony\Component\HttpFoundation\Response;
+
+class CustomRespondsToWebhook implements RespondsToWebhook
+{
+    public function respondToValidWebhook(Request $request, WebhookConfig $config): Response
+    {
+        return response()->json(['foo' => 'bar']);
+    }
+}