+++ tests/WebhookConfigTest.php
@@ -8,6 +8,7 @@ use Spatie\WebhookClient\SignatureValidator\DefaultSignatureValidator;
 use Spatie\WebhookClient\Tests\TestClasses\ProcessWebhookJobTestClass;
 use Spatie\WebhookClient\WebhookConfig;
 use Spatie\WebhookClient\WebhookProfile\ProcessEverythingWebhookProfile;
+use Spatie\WebhookClient\WebhookResponse\DefaultRespondsTo;
 
 class WebhookConfigTest extends TestCase
 {
@@ -24,7 +25,7 @@ class WebhookConfigTest extends TestCase
         $this->assertInstanceOf($configArray['signature_validator'], $webhookConfig->signatureValidator);
         $this->assertInstanceOf($configArray['webhook_profile'], $webhookConfig->webhookProfile);
         $this->assertEquals($configArray['webhook_model'], $webhookConfig->webhookModel);
-        $this->assertInstanceOf($configArray['process_webhook_job'], $webhookConfig->processWebhookJob);
+        $this->assertEquals($configArray['process_webhook_job'], $webhookConfig->processWebhookJobClass);
     }
 
     /** @test */
@@ -50,6 +51,26 @@ class WebhookConfigTest extends TestCase
     }
 
     /** @test */
+    public function it_validates_the_webhook_response()
+    {
+        $config = $this->getValidConfig();
+        $config['webhook_response'] = 'invalid-webhook-response';
+
+        $this->expectException(InvalidConfig::class);
+
+        new WebhookConfig($config);
+    }
+
+    /** @test */
+    public function it_uses_the_default_webhook_response_if_none_provided()
+    {
+        $config = $this->getValidConfig();
+        $config['webhook_response'] = null;
+
+        $this->assertInstanceOf(DefaultRespondsTo::class, (new WebhookConfig($config))->webhookResponse);
+    }
+
+    /** @test */
     public function it_validates_the_process_webhook_job()
     {
         $config = $this->getValidConfig();
@@ -68,6 +89,7 @@ class WebhookConfigTest extends TestCase
             'signature_header_name' => 'Signature',
             'signature_validator' => DefaultSignatureValidator::class,
             'webhook_profile' => ProcessEverythingWebhookProfile::class,
+            'webhook_response' => DefaultRespondsTo::class,
             'webhook_model' => WebhookCall::class,
             'process_webhook_job' => ProcessWebhookJobTestClass::class,
         ];